Basic setup of Material-UI
