+++ src/components/styling/MuiTheme.js
@@ -1,0 +1,39 @@
+import React from 'react'
+import { createMuiTheme } from '@material-ui/core/styles'
+
+const theme = React.useMemo(
+    () =>
+        createMuiTheme({
+            palette: {
+                primary: {
+                    light: '#e05425',
+                    main: '#a81e21'
+                },
+                secondary: {
+                    light: '#98ae4d',
+                    main: '#4f632c'
+                },
+                error: {
+                    main: '#f44336'
+                },
+                warning: {
+                    main: '#ff9800'
+                },
+                info: {
+                    main: '#2196f3'
+                },
+                success: {
+                    main: '#4caf50'
+                },
+
+                contrastThreshold: 3,
+                tonalOffset: 0.2
+            },
+
+            overrides: {
+            }
+        }),
+    []
+)
+
+export default theme
