+++ src/components/styling/Style.js
@@ -1,0 +1,7 @@
+import { makeStyles } from '@material-ui/core/styles'
+
+const useStyles = makeStyles((theme) => ({
+    
+}))
+
+export default useStyles
